+++ linear_actuator/LinearActuator.Core/SerialProtocol.cs
@@ -11,10 +11,10 @@ public static class SerialProtocol
 
     public static string FormatTargetCommand(ActuatorState state)
     {
-        int a1 = RequireValidTarget(state.A1Target, nameof(state.A1Target));
-        int a2 = RequireValidTarget(state.A2Target, nameof(state.A2Target));
-        int a3 = RequireValidTarget(state.A3Target, nameof(state.A3Target));
-        int a4 = RequireValidTarget(state.A4Target, nameof(state.A4Target));
+        double a1 = RequireValidTarget(state.A1Target, nameof(state.A1Target));
+        double a2 = RequireValidTarget(state.A2Target, nameof(state.A2Target));
+        double a3 = RequireValidTarget(state.A3Target, nameof(state.A3Target));
+        double a4 = RequireValidTarget(state.A4Target, nameof(state.A4Target));
 
         return $"T,{a1},{a2},{a3},{a4}\n";
     }
@@ -31,12 +31,12 @@ public static class SerialProtocol
         }
     }
 
-    public static bool IsValidTarget(int? value)
+    public static bool IsValidTarget(double? value)
     {
         return value is >= ActuatorConstants.MinTarget and <= ActuatorConstants.MaxTarget;
     }
 
-    private static int RequireValidTarget(int? value, string name)
+    private static double RequireValidTarget(double? value, string name)
     {
         if (!IsValidTarget(value))
         {
